Window Knocking
Knocking on windows is your main way of getting attention when you come to a department. It's a social feature that attracts other players to you.

This PR enhances that by making window knocking louder. The volume number isn't actually adjusted, but the ranges are increased and it has less falloff, especially the loud knocking (harm intent). This allows it to be heard from several rooms away, so you can bang on the front window of cargo until the merchant gets off his ass and comes to see who's outside.

In addition, this also allows you to knock on windoors, as well as normal windows. If you don't have access to a windoor when you click it, you'll knock on it instead. Same rules as on a normal window

If someone does this too much and becomes annoying, the ironhammer can come knock on their skull