Voidsuit and Magboot fixes (#1677)
This PR is branched off of #1676, and requires it. So review and merge that first. Otherwise it will appear to have a lot of junk files

This fixes a number of observed issues with magboots and voidsuits including #1658 and #1661

Modifiability of voidsuits is now standardised to work while held in hand or placed on the ground.
Attaching components to a voidsuit or toggling the helmet now has sounds.
Magboots massively refactored to function correctly and without any wierd behaviour.

Most notably, magboots were using the mob_can_equip, which is supposed to be a purely information gathering function, to actually affect the world. This was a dirty dirty hack, so its fixed