Dawn of the Dead (#1918)
A side project i've talked about too much and taken too long to finish.
This reworks respawning mechanics

Players now have three seperate respawn timers. The main one for playing as crew, an animal timer (for mice) and a drone timer.
Time of death for these types of mobs is tracked seperately. This means that you can play as one while waiting on the cooldown for another.

In addition, the initial death time of mice and drones is set such that an observer coming in from the main menu can immediately spawn as them, no waiting 5-10 minutes for the first time. The time between deaths remains though