Day of the Dead (#1920)
AKA, Respawning Tweaks, Part 2
This PR focuses on adding several means by which players can shorten the 30-minute crew respawn timer. There are three things that can affect it.

    Going into a cryobed at perfect health, will reduce respawn time by 20 minutes, allowing players to respawn as a new character fairly quickly.
    The health check is intended to approximately gauge the state of the round. Players should only be using cryo to respawn during peaceful times, when nothing is going on and they need to switch jobs. It shouldn't be used to escape from antags

In combination with this, the time taken for a cryobed to put someone into storage is reduced to ten minutes (previously 30) so players will be able to respawn at roughly the same time their body goes into storage.
also fixes #1683

    Your corpse being stored in a morgue drawer. This will reduce respawn time by 8 minutes, which is a relatively minor effect but every little helps, it mostly exists to encourage paramedics to get out there and recover corpses.
    In combination with this, morgue drawers are minorly refactored, a little code cleanup, more checks, and now only one mob and/or bodybag/coffin can be in each drawer.

    Your body being given a proper funeral. For now this specifically involves it being placed in a coffin and fired out of the mass driver. As long as a coffin containing a corpse leaves the ship zlevel, it will work. A funeral reduces respawn time by 15 minutes, a much stronger effect. This does not stack with the morgue drawer, overwriting it instead

In combination with this, I've fixed some odd behaviour with ztransitions. When someone goes off a map edge, they get sent to a random zlevel. In the longrun we will want to implement a proper solution to this involving the overmap. But for now, i've implemented a quick fix - if someone is leaving a ship zlevel, they will never arrive at a ship zlevel. This means that trash, corpses, and coffins fired out of the ship won't just loop back around and hit the hull

Finally some minor changes: The Become Mouse and Join as Drone verbs have been renamed, to Respawn as Mouse and Respawn as drone. This is so that they appear together in the ghost menu.
A duplicate of the Respawn verb has been added to the ghost menu to make it easier to find. It's just a wrapper that calls the base respawn verb

Note: This PR is branched off my previous one, Dawn of the Dead, so it contains extra changes that are covered there, and that should be handled first