Unreliable Protection (#2143)
First off, let me say, bay's armor system is terrible. IT is a poorly designed and even worse implemented mess. Requests to fix it are constant

Secondly, let me stress that I will replace it. Entirely, i'm going to tear it out and replace it with another system in the coming months. But that day may be next month or next year, it is not today.

This PR is a temporary measure for today. I'm not going to waste effort fixing the fundamentally bad design of the system, but i can at least make some sense of their terrible implementation

This PR Rebalances all the armor in the game, generally buffing it. Specifically targeting the weakest of the armors. The high end suits are -mostly fine- and have gotten only minor tweaking.

Why do this? Why is the bay system so bad? Here's a rough description of how it works:

Armor values are set between 0-100
When someone recieves a hit, they have a value% chance for the damage of that hit to be cut in half.
And IF that happens, they then get another value% chance to completely negate the hit

Why is this bad?

    Inconsistency. While 20 armor might sound like it'd reduce incoming damage by 20%, what it really does is give you an 80% chance to do absolutely nothing. In most fights, that 20 armor is not going to do anything for you and you'll feel like you're not wearing armor at all

    Scaling. The second check which potentially allows nullification, causes the effectiveness of armor values to snowball unpredictably out of control. I've done the math before, explored this quite extensively. Generally at values of 80 or more, someone is practically invincible, while below 40 it feels like it's doing nothing. It shoots up in a slightly exponential curve

The net effect is that the higher you go, the more each point of armor matters. At values above 50, it starts to become really meaningful, and changes should be made in small increments

    Wierd balance:
    The biggest thing wrong with the system, is that the people who wrote the values apparently didn't understand points 1 and 2. Bay armor has a couple of specifically notable quirks that i've fixed here:

3a: Station armor and lasers
For reasons beyond my ken, and which certainly don't apply here, bay wanted station security to use lasers, and antags to use bullets. And they wanted these two to kill each other really easily. The result of this is that default sec/IH armor had okayish laser resist but completely worthless bullet resistance. The values were hardly balanced even in relation to each other, the underlying goal of them is questionable, and on eris, where ballistic weapons are easily available, it simply makes no sense
This is fixed up and IH armor now has 40 in both laser and bullet resist

3b. Specialised armors
There are three specialised things in bay; bulletproof, ablative and riot armor. They're supposed to offer high resist against one type (bullets, lasers and melee) while being weak against the other two types.
This is actually a pretty good idea and I approve of it. These armors are the strongest in the game in their respective fields, and for now i'm letting that continue.
But the weak points are a cause for concern, they gave them a value of only 15, which is worthless. I've buffed these to be only slightly weaker than the default IH armor, so there's still a meaningful choice (specialisation vs generalisation) but it doesnt quite feel like you're wearing paper

Anyways i've written a lot already. Bottom line, this PR is a much needed and requested temporary fix
Armor system due for future replacement