Adds the RXD, a ghetto crossbow made from RCDs (#2411)
The RXD is a RCD with crossbow parts welded onto it. It flashforges bolts using matter units when its string is drawn back. Each bolt costs 10 matter units to forge, and flashforged bolts can be reclaimed for 5 matter units each. It cannot be manually loaded with bolts, nor can its bolts be removed except when firing.