Preacher Ability: Atonement
Adds a new ability for the Preacher, 'Atonement'.

Atonement is essentially the Inquisitor's power 'Penance', but costs more to use which stops the Preacher from being able to completely disable someone with it.

It will however greatly slow the person down and they may pass out for a few short moments if cast twice in a row (the energy cost is set to make it require at least a few moments before being able to cast again).

It's intended to be used to punish people for not obeying Church doctrine and such.