Implemented Suggestions
1. Renamed IH energy armor to IH ablative armor and added a random pouch crate under operations.

2. Fixed the price for the hardsuit crates - had meant to make it higher but I apparently forgot, my b. It's 5k for a random one that can be damaged, now.

3. Added new IH guns to surplus crates.

4. Raised cost of plasteel and assault crate, but kept weapons the same. Makes more sense to me that asters would offer paramilitary gear, like the wt550 and the ak, to private buyers and keep their premium stuff for contracts like IH. I also like splitting the aesthetic so IH uses IH guns and everyone else uses other guns.

As to it being a buff, yes, I agree that cargo comes out stronger than they did before. I do believe that the state they were in previously was not intentional, but simply due to lack of work being done on keeping their orderable stock up to date with our code.