Manual Redux (#3523)
Requires eventual replacement of existing books with map changes, but not in this PR.

Properly connects a new set of wiki manuals, which will connect to the wiki instead of the broken system the manuals currently use. Includes language support, and the wiki should connect to the Russian version of the page if it's on the Russian server and vice-versa.

It just works.