Augments QoL
Reworks engineering and surgical toolset implants to greatly improve usbaility. They no longer act like built-in toolkits; instead, they function as omnitools that can perform any of the necessary functions of the tools that were in the implant before. The implants also contain an analyzer and T-ray scanner, and a health scanner, respectively.