Augments QoL Changes
Reworks engineering and surgical toolset implants to greatly improve usbaility. They no longer act like built-in toolkits; instead, they function as omnitools that can perform any of the necessary functions of the tools that were in the implant before. The implants also contain an analyzer and T-ray scanner, and a health scanner, respectively. Additionally, added two new sprites to tools.dmi, one is just an edited version of the onestar medical omnitool, used for the surgical omnitool, and the other was provided to me by Holvyr (idr what it's from).