Further augments QoL changes (#4223)
* Augments QoL Changes

Reworks engineering and surgical toolset implants to greatly improve usbaility. They no longer act like built-in toolkits; instead, they function as omnitools that can perform any of the necessary functions of the tools that were in the implant before. The implants also contain an analyzer and T-ray scanner, and a health scanner, respectively. Additionally, added two new sprites to tools.dmi, one is just an edited version of the onestar medical omnitool, used for the surgical omnitool, and the other was provided to me by Holvyr (idr what it's from).

* Deleted accidental additions

* Augments QoL Changes

Reworks engineering and surgical toolset implants to greatly improve usability. They no longer act like built-in toolkits; instead, they function as omnitools that can perform any of the necessary functions of the tools that were in the implant before. The implants also contain an analyzer and T-ray scanner, and a health scanner, respectively. Additionally, added two new sprites to tools.dmi, one is just an edited version of the onestar medical omnitool, used for the surgical omnitool, and the other was provided to me by Holvyr (idr what it's from).

* Restoring accidental pricetag deletion

* Descriptions / comments updated

* Removed unnecessary whitespace

* Removed shoveling from engi implant

* General cleanup and tweaking

- Made the surgical omnitool a sharp, edged weapon with PAINFUL force
- Made the engineering omnitool a blunt weapon with DANGEROUS force
- Removed the glow from the engineering omnitool (kind of obnoxious)
- Greatly buffed degradation for both implants but nerfed workspeed to compensate. Also increased power use for both weapons. 
- Removed matter vars from the implants so people don't accidentally insert them into the autolathe
- Made the engi implant conductive

* Create eyes.dm

* Delete eyes.dm

* Update _external.dm

* Aug QoL changes

Removed scanners from the surgical and engineering augs and made them deploy instantly without a popup menu.

* Fixing things

* Update misc.dm

* moved new omnitool augs out of the multitool folder

* fixed filepaths in the dme