Reduces availability of Harpoon and Teleporter module.
It is odd how a device apparently so rare, and coming from a now extinct producer comes in such a great abundance. The Teleporter module never was once obtainable, and was designed for use by a Ninja antag on Bay who was meant to be peaceful. Not empty out the entire ship into space, as does every Traitor who discovers this effective function.