Cyberplants give positive sanity value to area.
Cyberplants give positive sanity value to the area that they are located in, much needed for the ship's public areas, and gives a mechanical depth to decor. For reference, it takes 5 cyberplants to reach the same sanity value as a single oddity, so this will mostly benefit large public areas.