Colt Rebalance/Nerf
Colt is currently by far the best pistol in the game despite being the oldest design. It's also cheaper than pistols that are much worse than it.

Since it is such an old design, wanted to make it more like the novakovic. Doesn't cost plasteel to manufacture and is sold in the vendor cheaply while being respectable but outclassed by more modern guns.