Borer rebalance and flavor text change.
Switched the drug that makes the borer docile to mindbreaker, increased the time needed to infect a cruciformed player, increased the cooldown time of the paralysis ability and the duration of the effects. Lastly the flavor test can been changed to be less psychic space slug and more organic poison dart shooting parasite.