Borer rebalance, flavor text changed and devour re-enabled
Made the delay timer for infecting a NT take longer, increased the cooldown timer and effect duration for the paralyse ability, also changed its flavor text to lean more into a biological parasite that doesn't have pyschic abilities. Re-enabled the devour ability which will allow the borer to convert the dead host into a husk if the borer was controlling it when it died. Switched the docile reagent from sugar to mindbreaker.