Update Sleeping
Slightly increases the halloss healing across the board when resting/incapacitated as a stop-gap until more measures to bring pain into line are taken, also adds a base level of brute, tox and burn healing. Oxygen healing is excluded as oxygen is often used as a death timer, and sleeping should not be a get out of death free card in dire situations, but a way of recovering from injuries that might just be enough to tip someone into dying under the current system. 

Numbers could be tweaked downward but I feel like this is a good and decent base.