Scavenger nerf
Not a full revert, but bullet resistance has been compromised at 35 in line with flakvests, while energy resistance was reduced to 25 as Scavengers won't have to expect facing many laser-wielding opponents.
The built-in light was a little silly, so it has been removed for the time being. If the sprite and recipe is updated, then it may be readded.