Fixes runtime when Iriska eats untouched food
It seems Iriska breaks when she eats food that wasn't dirtied by the hands of a living being. A good example of such food is freshly spawned food that was not yet interacted with.

Iriska will check if the last person to touch them (such person is null) is not in the despised list (the list does not contain Mr Null), then attempts to tolerate them.

This of course, causes her to runtime. A check is made to make sure the person is indeed valid.

Despise can also runtime this way, likely from atmossed items, or some other projectiles not launched by living mobs. It receives this check as well.