Factorio, but with plants (Alchemy Reigns) (#8100)
* Add RealEggPlant, Protein

Reintroduces Real Egg Plant, which I found referenced as a mutation of eggplants but the rest was empty. Filled it in, gave it yolk as an additional reagent, and added it's seed packet data and put it into the Curious Seeds of the NT Vendor. Also introduced animal protein as a reagent to blood tomatoes.

* Update seed_datums.dm

Adapt funnier name

* Update nt_cruisers.dm

Adapt funnier name