Sorter Room Disposal Chute Fix, Shipping Supplies (General Map PR) (#8327)
* Fix Sorter Chute

Disposal Chute leading to sorter somehow never had a trunk under it. Really ever. How in tarnation was this ever missed?

* Add Stairwell From Club to Gym

Use old NT 'janitor supply' as a hodgepodge stairwell up to the center of the gym.

* Add Tagger/Wrap around map

Added Destination Tagger and Package Wrap to Heads of Department rooms, and any existing instance of EFTPOS.

* Neothology EFTPOS

Made and added Neotheology EFTPOS to both the church stall, and the Preacher's office.

* Add Tagger/Wrap to Library

Makes a second spot that public Destination Tagger and Package Wrap are available to those outside departments or non-heads.

* Remove Club Stairs

Removes the "new" stairs from club to gym.

* Corrected NT Windoor Access

When I added the windoor to the agro room, I just made a copy of the NT shop windoor, assuming it had access proper. It didn't, so I added "Crematorium" access to the shop (which is limited to all clergy), and Agrolyte access to the agro windoor.

* Map Spare Pulsar Power Bridge

Added a spare Pulsar Power Bridge circuit to secured technical storage, so it exists as more than maint loot.