Tiers, Bugfix
Created and adjusted quality tiers, and added it as a multiplier to sanity_gain on each bite.

Found bug on get_sanity_gain that only obtained "nutriment" from food to measure with, resulting in food such as meat steak and others that contained nutriment children to be left out.

Defaulted timers to "infinite" (0) instead of 1.